Add stage/transfer API tests to CI framework
Creates lower-level tests (t/ci/800-stage-tests.sh) to do a proof of
concept of the transfer API. Also creates a higher-level file
(t/ci/RUN_CI_STAGE_TESTS.sh) that runs the lower-level tests over a
sweep of file sizes and server configurations.

Added a parallel transfer test to t/ci/800-stage-tests.sh as well.
It is currently being skipped until the parallel transfer logic is
fixed.

Fixes a bug in serial transfer logic causing the destination file
to be created as read-only, resulting in a failed transfer.

Update transfer docs to account for the various methods.

+++ t/ci/990-stop-server.sh
@@ -5,16 +5,22 @@
 
 test_description="Stopping the UnifyFS server"
 
-while [[ $# -gt 0 ]]
+DONE_TESTING="1"
+
+for arg in "$@"
 do
-    case $1 in
+    case $arg in
         -h|--help)
             ci_dir=$(dirname "$(readlink -fm $BASH_SOURCE)")
             $ci_dir/001-setup.sh -h
             exit
             ;;
+        --allow-restart)
+            DONE_TESTING="0"
+            echo "Stop servers but allow restart and additional testing"
+            ;;
         *)
-            echo "usage ./990-stop-server.sh -h|--help"
+            echo "usage ./990-stop-server.sh -h|--help|--allow-restart"
             exit 1
             ;;
     esac
@@ -34,32 +40,42 @@ test_expect_success "verify unifyfsd has stopped" '
     test_must_fail process_is_running unifyfsd 10
 '
 
-# If UNIFYFS_MOUNTPOINT is an existing dir, verify that is it empty
-test_expect_success REAL_MP "Verify UNIFYFS_MOUNTPOINT ($UNIFYFS_MP) is empty" '
-    test_dir_is_empty $UNIFYFS_MP
-'
 
-# Cleanup posix mountpoint
-test_expect_success POSIX "Cleanup UNIFYFS_CI_POSIX_MP: $UNIFYFS_CI_POSIX_MP" '
-    rm -rf $UNIFYFS_CI_POSIX_MP/*posix*
-'
+if [ "$DONE_TESTING" -eq "1" ]; then
 
-# cleanup_hosts
-test_expect_success PDSH,CLEAN "Cleanup hosts" '
-    cleanup_hosts
-'
-# Remove trap
-# If any tests failed, the suite will exit with 1 which will trigger the trap.
-# Since the hosts were already cleaned at this point, can remove trap to prevent
-# cleanup_hosts from being called again.
-trap - EXIT
 
-# end here if running tests individually
-if [[ -z $full_run ]]; then
-    ( test_done; )
-    test_exit_code=$?
+    # If UNIFYFS_MOUNTPOINT is an existing dir, verify that is it empty
+    test_expect_success REAL_MP \
+      "Verify UNIFYFS_MOUNTPOINT ($UNIFYFS_MP) is empty" '
+        test_dir_is_empty $UNIFYFS_MP
+    '
 
-    cd "$(dirname "$SHARNESS_TRASH_DIRECTORY")"
+    # Cleanup posix mountpoint
+    test_expect_success POSIX \
+      "Cleanup UNIFYFS_CI_POSIX_MP: $UNIFYFS_CI_POSIX_MP" '
+        rm -rf $UNIFYFS_CI_POSIX_MP/*posix*
+    '
 
-    return $test_exit_code
-fi
+    # cleanup_hosts
+    test_expect_success PDSH,CLEAN "Cleanup hosts" '
+        cleanup_hosts
+    '
+
+    # Remove trap
+    # If any tests failed, the suite will exit with 1 which
+    # will trigger the trap.  Since the hosts were already
+    # cleaned at this point, can remove trap to prevent
+    # cleanup_hosts from being called again.
+    trap - EXIT
+
+    # end here if running tests individually
+    if [[ -z $UNIFYFS_CI_TESTS_FULL_RUN ]]; then
+        ( test_done; )
+        test_exit_code=$?
+
+        cd "$(dirname "$SHARNESS_TRASH_DIRECTORY")"
+
+        return $test_exit_code
+    fi
+
+fi # if [ $DONE_TESTING -eq "1" ]]; then
